Refactor color retrieval in BaseDiscourseNodeUtil: replace manual color logic with getDiscourseNodeColors utility function for improved maintainability and clarity.
